+++ vermar/src/pages/new.tsx
@@ -8,7 +8,7 @@ import { Button } from "../components/ui/Button";
 import { Form } from "../components/ui/Form";
 import { Input } from "../components/ui/Input";
 import { Markdown } from "../components/ui/Markdown";
-import { TextArea } from "../components/ui/Textarea";
+import { MarkdownEditor } from "../components/ui/MarkdownEditor";
 import { PostInput, useCreatePostMutation } from "../generated/graphql";
 import { useAppPersistStore } from "../store/useAppStore";
 import withApollo from "../utils/withApollo";
@@ -52,7 +52,7 @@ const New = () => {
                 <h1 className="mb-4">Create a post</h1>
                 <Form form={form} onSubmit={onSubmit}>
                   <Input label="Title" {...form.register("title")} />
-                  <TextArea
+                  <MarkdownEditor
                     label="Content"
                     {...form.register("content")}
                     count={source.length}
